Remove legacy tests and fix up some that I broke
